Cargo.toml: Add release profile for smaller size and overflow checks
The release binary should be as small as possible without impact. Since
this is new code we also want the release binary to still have overflow
checks instead of silently eating them.
Add a release profile that strips the debug symbols, uses one codegen
unit to get similar gains as thin LTO (which we seem to have cross-
compiling problems with), and enable overflow checks.