Switch to reqwest::blocking and use a File object
The use of async for the client was motivated by doing hashing while
downloading and ideally everything else too so it would extract to disk.
However, only the hashing was done and all other operations including
and additional hash step were done on the file.
Since the use of async brought more problems while not having a real
benefit, switch to a blocking mode. Instead of opening the file and
passing a write to the download function we now open the file inside
the download function which will make it easier to retry downloads.