Use explicit offsets for reading and writing
Currently the order in which the functions are called matters because
they advance the offset. Some had seeks but not all.
Use explicit read and write offsets to make the code more robust. Also,
we have to make use of the start_block to calculate the destination even
if we didn't ran into a problem yet because the way we generate the
payload was giving the order we happened to write out.