omaha: remove Uuid type and provide unit tests for xml read/write
This was documenting as existing to provide serialisation of the underlying
inner value of type uuid::Uuid with surrounding braces. Whenever uuids are
used throughout the codebase I image they are used just as uuid::Uuids,
therefore we don't need a type just for (de)seriailsation purposes -- just
provide a `with = ` attribute for XmlRead and XmlWrite.

Also provide here preliminary unit tests for reading/writing some omaha
structs using the base uuid::Uuid type to prove this is sufficient.

Signed-off-by: james-parky <72101291+james-parky@users.noreply.github.com>