omaha: remove type.rs and therefore the FileSize struct
It doesn't appear to have a need to exist; using a FileSize in places that
could just be a usize doesn't provide any extra type-safety, and doesn't
really provide any extra semantic meaning to a dev. Wherever it was used it
was just being converted back and forth from a usize anyway. Furthermore,
there is no special (de)serialisation logic that is required for it -- it
is also just a usize in the XML.

Signed-off-by: james-parky <72101291+james-parky@users.noreply.github.com>