download: Fix bug of pkg_verified path extension
Before PR https://github.com/flatcar/ue-rs/pull/87, target_filename was
almost always `None`, which had covered an issue of setting pkg_verified.
In case of target_filename being a valid String, for example
"oem-azure.gz", pkg_verified ended up having an extension `.gz`, which
is not expected.

Now that the PR was merged, DownloadVerify module gets created with
.target_filename(args.target_filename.unwrap_or(TARGET_FILENAME_DEFAULT.into())),
where TARGET_FILENAME_DEFAULT is with ".gz" extension. That uncovered
the hidden issue.

Fix that with explicitly setting extension ".raw" for both a valid
String and None for the input string.

Signed-off-by: Dongsu Park <dongsu@dpark.io>